Improve source code layout.
The design wouldn't have worked if there were several files because it
would spill over into the next line. This was reworked into a dropdown
for student files, a dropdown for instructor files, and then the diff
file by itself. The JS was then updated accordingly to match the new
HTML structure.