Remove prettydiff library...
Spent a good chunk of time learning how to use prettydiff. Very
frustrating to use. The last straw was discovering that it overwrites
styles in tags (like h1, h2, and many more) without being specific to a
class or id. Very, very bad and impossible to maintain. Switching to
jsdifflib!